requirements.txt
graia-ariadne>=0.10.3
Pillow>=9.3.0
revChatGPT==1.2.1
openai
tiktoken
toml
aiohttp
asyncio